Change AU grid to have extra padding on the South and West sides
Move tile `0,0` 5 more tiles to the West and 4 more tiles to the South. We now
have a padding of 5 tiles outside of the valid range for both X and Y, used to
be 1 extra tile along Y dimension only.

This is to support Sentinel 2 data holdings and allow some safety margin for the
future.